Specify UTC time zone in example
ISO:8601 allows to specify the timezone to unambiguously mark it as UTC via
appending a "Z". See e.g. 5.3.3 in
http://www.cs.utsa.edu/~wagner/date/8601.pdf.

Use that here to give future users a nice example.

The default timezone for the python module iso8601 is UTC, so this is
ambiguity only exists for other programming languages.

Signed-off-by: Thomas Braun <thomas.braun@byte-physics.de>